Separate buffers and the buffer pool into their own modules.
