+++ core/src/plugin/event.rs
@@ -59,3 +59,7 @@ impl ProcEvent {
         }
     }
 }
+
+pub trait CoreEventExt {}
+
+impl<'a> CoreEventExt for CoreEventSpace<'a> {}
